Improved inference, use nore generic collections
Got some ideas from Scalaz, which made the generic collections even more
high-level. There are two cases now, one for traversables with one type
parameter and one for maps.

By using a nice trick in MapReduce, we removed the redundancy of the
element type (both in de function definition and for the map reduce
call). It is now inferred from the map function, only the expected return
collection has to be specified.